+++ api/src/main/java/dev/lambdaurora/lambdynlights/api/entity/luminance/MagmaCubeLuminance.java
@@ -11,19 +11,24 @@ package dev.lambdaurora.lambdynlights.api.entity.luminance;
 
 import dev.lambdaurora.lambdynlights.api.item.ItemLightSourceManager;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.monster.cubemob.AbstractCubeMob;
-import net.minecraft.world.entity.monster.cubemob.MagmaCube;
 import org.jetbrains.annotations.Range;
 
 /**
  * Provides the luminance value of a magma cube.
  *
  * @author LambdAurora
- * @version 4.1.0
+ * @version 4.11.0
  * @since 4.1.0
+ * @deprecated Prefer {@link CubeMobSquishLuminance} instead.
  */
+@Deprecated
 public final class MagmaCubeLuminance implements EntityLuminance {
 	public static final MagmaCubeLuminance INSTANCE = new MagmaCubeLuminance();
+	private final CubeMobSquishLuminance actual = new CubeMobSquishLuminance(
+			EntityLuminance.of(8),
+			EntityLuminance.of(11),
+			0.6f
+	);
 
 	private MagmaCubeLuminance() {}
 
@@ -34,10 +39,6 @@ public final class MagmaCubeLuminance implements EntityLuminance {
 
 	@Override
 	public @Range(from = 0, to = 15) int getLuminance(ItemLightSourceManager itemLightSourceManager, Entity entity) {
-		if (entity instanceof AbstractCubeMob magmaCube) {
-			return (magmaCube.squish > 0.6) ? 11 : 8;
-		}
-
-		return 0;
+		return this.actual.getLuminance(itemLightSourceManager, entity);
 	}
 }